+++ test/html_renderer_test.dart
@@ -26,7 +26,9 @@ void main() {
         withDefaultInlineSyntaxes: false,
         encodeHtml: false,
         blockSyntaxes: [const HorizontalRuleSyntax()],
-        inlineSyntaxes: [TagSyntax(r'\*+', requiresDelimiterRun: true)],
+        inlineSyntaxes: [
+          EmphasisSyntax.asterisk(),
+        ],
       );
 
       expect(